# auth-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_likhon22_ad-system_auth-service_internal_domain.Role:
    enum:
    - advertiser
    - publisher
    - admin
    type: string
    x-enum-varnames:
    - RoleAdvertiser
    - RolePublisher
    - RoleAdmin
  github_com_likhon22_ad-system_auth-service_internal_domain.Status:
    enum:
    - active
    - suspended
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusSuspended
  github_com_likhon22_ad-system_auth-service_internal_domain.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      passwordHash:
        type: string
      role:
        $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_domain.Role'
      status:
        $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_domain.Status'
      updatedAt:
        type: string
    type: object
  github_com_likhon22_ad-system_auth-service_internal_utils.ErrorResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  ? github_com_likhon22_ad-system_auth-service_internal_utils.Response-github_com_likhon22_ad-system_auth-service_internal_domain_User
  : properties:
      data:
        $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_domain.User'
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_adapter_inbound_http_handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_adapter_inbound_http_handler.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - advertiser
        - publisher
        type: string
    required:
    - email
    - password
    - role
    type: object
host: localhost:5000
info:
  contact: {}
  description: Authentication service for the Ad System
  title: Auth Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user account with email, password
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_inbound_http_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.Response-github_com_likhon22_ad-system_auth-service_internal_domain_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.ErrorResponse'
      summary: Login a new user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email, password and role
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_inbound_http_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.Response-github_com_likhon22_ad-system_auth-service_internal_domain_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_likhon22_ad-system_auth-service_internal_utils.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"